+++ src/pages/location/index.tsx
@@ -1,10 +1,10 @@
-import NavBar from '@components/common/NavBar';
+import NavBar from '@components/navbar/NavBar';
 import PageLayout from '@components/layout/PageLayout';
 
 const Location = () => {
   return (
     <PageLayout>
-      <NavBar area="지도 지역명" isSearch={true} isRight={true} />
+      <NavBar area="지도 지역명" isRight={true} />
       Location
     </PageLayout>
   );
